requirements.txt
fastapi
uvicorn
imagehash
opencv-python
Pillow
pytesseract
simhash
openai-whisper
ffmpeg-python
python-magic
pymongo
PyMuPDF
python-multipart
streamlit
matplotlib
transformers
scikit-learn
sentence-transformers
chromadb
fpdf